Basic detector for DeLTA2.
Runs the *tracking* process, not only the detection.
Delta (if v>3) runs and creates the netCDF output, but crashes
because it cannot find FFMPEG. No big deal.
Also: it does not parse the output yet. Which will be tricky, because
DeLTA2 is actually a tracking algorithm, not only a detection
algorithm.